+++ src/load.rs
@@ -4,6 +4,9 @@ use std::fs::File;
 use std::path::Path;
 use protos::master::Game;
 use screen::Interfaceable;
+use protobuf::CodedOutputStream;
+use protobuf::Message;
+use io::write_protobuf;
 
 #[allow(unused_mut)]
 pub fn load<I: Interfaceable>(
@@ -12,13 +15,6 @@ pub fn load<I: Interfaceable>(
 ) -> Result<Game, String> {
     src.print("I see you've been a guest with us before.");
     src.print(&format!("Welcome back {}.", game.get_name()));
-    return Ok(game);
-}
-
-pub fn save<I: Interfaceable>(
-    src: &mut I,
-    mut game: Game,
-) -> Result<Game, String> {
     return Ok(game);
 }
 
@@ -31,13 +27,5 @@ pub fn new<I: Interfaceable>(
 
     let name = src.prompt();
     game.set_name(name.clone());
-
-    src.print(&format!(
-        "Hello {}, please enter the name of the new save file:",
-        &name
-    ));
-    let path = src.prompt();
-
-    File::create(&Path::new(&path)).unwrap();
-    return Ok(game);
+    return write_protobuf(src, game);
 }
